Modified access paths and redirecting
Modified any access paths - to the work files which are needed for
XMLRecordCreatorTests (now it is more general) and also to the installed
texlive library in the system (also more general). And modified access
in situations when anybody want to show information which does not exist
- now is redirected to the intro page.

+++ java/CV-Generator/src/java/cz/muni/fi/server/NewCVCreation.java
@@ -90,12 +90,26 @@ public class NewCVCreation extends HttpServlet {
                 xslt.transformToTex("xml_to_tex.xslt", person.getDateHash()+".xml", 
                                     person.getDateHash()+".tex",contextPath);
                 File texFile = new File(contextPath+"/pdf_database",person.getDateHash());
-                PDFfromLatexBuilder pflb = new PDFfromLatexBuilder("C:\\texlive\\2013\\bin\\win32\\");
+                String pathToTeXLive = null;
+                File[] roots = File.listRoots();
+                for(File root: roots){
+                    String path = root.getName()+"/texlive/2013/bin/win32/";
+                    File f = new File(path);
+                    if(f.exists()){
+                        pathToTeXLive = path;
+                        break;
+                    }
+                }
+                PDFfromLatexBuilder pflb = new PDFfromLatexBuilder(pathToTeXLive);
                 pflb.createPDF(texFile);
                 File fOutToDelete = new File(texFile+".out");
                 File fTexToDelete = new File(texFile+".tex");
-                fOutToDelete.delete();
-                fTexToDelete.delete();
+                if(fOutToDelete.exists()){
+                    fOutToDelete.delete();
+                }
+                if(fTexToDelete.exists()){
+                    fTexToDelete.delete();
+                }
                 File invalidFile = new File(contextPath,"invalid.xml");
                 if(invalidFile.exists()){
                     invalidFile.delete();
